+++ packages/_private/demo-utilities/src/stubs/chart.ts
@@ -1,4 +1,5 @@
-import type { ECConfig, LineSeriesOption } from '@synergy-design-system/components/components/chart/types.js';
+import type { ECConfig, GaugeSeriesPresetOptions, LineSeriesOption } from '@synergy-design-system/components/components/chart/types.js';
+import { formatter } from '@synergy-design-system/components/components/chart/index.js';
 
 export const lineChartSeriesData: LineSeriesOption[] = [
   { data: [150, 230, 224, 218, 135, 147, 260], name: 'Visits', type: 'line' },
@@ -13,4 +14,24 @@ export const generalChartConfig: ECConfig = {
 export const lineChartConfigObject: ECConfig = {
   series: lineChartSeriesData,
   ...generalChartConfig,
+};
+
+export const gaugeChartConfigObject: GaugeSeriesPresetOptions = {
+  formatter: {
+    max: formatter.unitFormatter('°C'),
+    min: formatter.unitFormatter('°C'),
+    value: formatter.unitFormatter('°C'),
+  },
+  max: 120,
+  min: 10,
+  sections: {
+    boundaries: [10, 40, 70, 120],
+    show: true,
+  },
+  trend: {
+    direction: 'down',
+    show: true,
+    value: '6.5%',
+  },
+  value: 82,
 };