+++ packages/metadata/data/layers/full/tokens/charts/js/index.js
@@ -1,5 +1,5 @@
 /**
- * @synergy-design-system/tokens version 3.23.0
+ * @synergy-design-system/tokens version 3.24.0
  * SICK Global UX Foundation
  */
 
@@ -697,3 +697,8 @@ export const SynChartSequentialStatusWarning100 = 'var(--syn-chart-sequential-st
  * @type {string}
  */
 export const SynChartTooltipColor = 'var(--syn-chart-tooltip-color)';
+
+/**
+ * @type {string}
+ */
+export const SynChartTrackColor = 'var(--syn-chart-track-color)';